// src/shared/lib/types/Product.ts
export type Product = {
  id: number
  title: string
  description: string
  price: number
  discountPercentage: number
  rating: number
  stock: number
  brand: string
  category: string
  thumbnail: string
  images: string[]
}

export type ProductsResponse = {
  json(): unknown
  products: Product[]
  total: number
  skip: number
  limit: number
}